Allow not using redis for setup step.

+++ magento2-nginx/fpm-7.0/usr/local/share/magento2/magento_functions.sh
@@ -74,8 +74,10 @@ function do_magento_move_compiled_assets_away_from_codebase() {
 
 function do_magento_setup_upgrade() {
   rm -rf var/generation/*
-  redis-cli -h "$REDIS_HOST" -p "$REDIS_HOST_PORT" -n "$MAGENTO_REDIS_CACHE_DATABASE" "FLUSHDB"
-  redis-cli -h "$REDIS_HOST" -p "$REDIS_HOST_PORT" -n "$MAGENTO_REDIS_FULL_PAGE_CACHE_DATABASE" "FLUSHDB"
+  if [ "$MAGENTO_USE_REDIS" = "true" ]; then
+    redis-cli -h "$REDIS_HOST" -p "$REDIS_HOST_PORT" -n "$MAGENTO_REDIS_CACHE_DATABASE" "FLUSHDB"
+    redis-cli -h "$REDIS_HOST" -p "$REDIS_HOST_PORT" -n "$MAGENTO_REDIS_FULL_PAGE_CACHE_DATABASE" "FLUSHDB"
+  fi
   as_code_owner "bin/magento setup:upgrade"
 }
